+++ CHANGELOG.md
@@ -20,11 +20,12 @@
 - **lsp:** added yamlls custom schemas
 - **lsp:** disabled open float in favor of lsp_lines
 - **lsp-insstall:** eslint
-- **lsp_lines:** added lsp_lines
 - **lsp_lines:** changed packer declaration using as
+- **lsp_lines:** added lsp_lines
 - **markdown-preview:** added markdown preview
 - **neovim-dap:** added command to easily debug neovim itself
 - **nvim-lspinstall:** migrated to nvim lsp install
+- **pounce:** s key now uses pouce. the rest still uses vim sneak
 - **pretty-fold:** added pretty fold
 - **pretty-fold:** set fold next max to 1
 - **scrollbar:** added scrollbar
@@ -35,9 +36,9 @@
 - **telescope:** added telescope env
 - **treesitter:** foldexprt now to treesitter
 - **types:** added types for lua
+- **which-key:** shortcut to show mappings has changed from '?' to 'g?'
 - **which-key:** added dofile on <leader>pR
 - **which-key:** return back to folke's version
-- **which-key:** shortcut to show mappings has changed from '?' to 'g?'
 - **windline:** added lsp signature
 
 ### Fix